+++ RELEASES.rst
@@ -2465,6 +2465,11 @@ Planned
   necessary because all free operations decrement the voluntary GC counter and
   all allocs/reallocs check for voluntary GC (GH-1355)
 
+* Use a 32-bit refcount field by default (even on 64-bit systems) which saves
+  8 bytes for each heap object and can only wrap if the Duktape heap is
+  larger than 64GB; disable DUK_USE_REFCOUNT32 to use size_t for refcounts
+  (GH-1399, GH-1401)
+
 * Fix a garbage collection bug where a finalizer triggered by mark-and-sweep
   could cause a recursive entry into mark-and-sweep (leading to memory unsafe
   behavior) if the voluntary GC trigger counter dropped to zero during